bilbo contribution!! Mummi contribution!! veriverse contribution!! LCTV contribution!! catherinacortega, fuel, jegas, xmetrix, epryka, zzzel, markus, bananastriker, reddevil, gosu and magrib contribution!!
Thank you Carlos!!

Thank you Mummi!!

Carol contribution!!

at bilbo at veriverse!!
i7 6700 16GB RAM SSD OCZ 120GB SSD 240GB HD 2TB Windows 10 Pro

+++ axis/rdpclient/client/Windows/wf_client.cpp
@@ -325,7 +325,7 @@ CLASS_DECL_EXPORT BOOL wf_pre_connect(freerdp* instance)
 		return FALSE;
 	}
 
-	freerdp_set_param_uint32(settings, FreeRDP_KeyboardLayout, (int) GetKeyboardLayout(0) & 0x0000FFFF);
+	freerdp_set_param_uint32(settings, FreeRDP_KeyboardLayout, (size_t) GetKeyboardLayout(0) & 0x0000FFFF);
 
 	PubSub_SubscribeChannelConnected(instance->context->pubSub,
 		(pChannelConnectedEventHandler) wf_OnChannelConnectedEventHandler);
